# openapi/oh.yaml
openapi: 3.1.0
info:
  title: Open Hospital API Documentation
  description: Open Hospital API Documentation
  contact:
    name: ApiInfo.DEFAULT_CONTACT
  license:
    name: GPL-3.0 license
    url: https://github.com/informatici/openhospital-api?tab=GPL-3.0-1-ov-file#readme
  version: 0.1.0
servers:
- url: http://API_HOST:API_PORT
security:
- bearerAuth: []
...
paths:
...
  /patients/{code}:
    get:
      tags:
      - Patients
      operationId: getPatient
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDTO"
      security:
      - bearerAuth: []
    put:
      tags:
      - Patients
      operationId: updatePatient
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDTO"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Patients
      operationId: deletePatient
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
      - bearerAuth: []
  /patientconsensus/{patientId}:
    get:
      tags:
      - Patient Consensus
      operationId: getPatientConsensus
      parameters:
      - name: patientId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientConsensusDTO"
      security:
      - bearerAuth: []
    put:
      tags:
      - Patient Consensus
      operationId: updatePatientConsensus
      parameters:
      - name: patientId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientConsensusDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientConsensusDTO"
      security:
      - bearerAuth: []
...
  /patients:
    get:
      tags:
      - Patients
      operationId: getPatients
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 80
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagePatientDTO"
      security:
      - bearerAuth: []
    post:
      tags:
      - Patients
      operationId: newPatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDTO"
      security:
      - bearerAuth: []
  /patients/by-codes:
    post:
      tags:
      - Patients
      operationId: getPatientByCodes
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int32
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientDTO"
      security:
      - bearerAuth: []
...
  /visits/patient/{patID}:
    get:
      tags:
      - Visit
      operationId: getVisit
      parameters:
      - name: patID
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VisitDTO"
      security:
      - bearerAuth: []
...
  /patients/search:
    get:
      tags:
      - Patients
      operationId: searchPatient
      parameters:
      - name: firstName
        in: query
        required: false
        schema:
          type: string
          default: ""
      - name: secondName
        in: query
        required: false
        schema:
          type: string
          default: ""
      - name: birthDate
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: address
        in: query
        required: false
        schema:
          type: string
          default: ""
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientDTO"
      security:
      - bearerAuth: []
  /patients/nextcode:
    get:
      tags:
      - Patients
      operationId: getPatientNextCode
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
      security:
      - bearerAuth: []
  /patients/merge:
    get:
      tags:
      - Patients
      operationId: mergePatients
      parameters:
      - name: mergedcode
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: code2
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
      - bearerAuth: []
  /patients/cities:
    get:
      tags:
      - Patients
      operationId: getPatientCities
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      security:
      - bearerAuth: []
  /patients/all:
    get:
      tags:
      - Patients
      operationId: getPatientAll
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDTO"
      security:
      - bearerAuth: []
...
  /opds/patient/{pcode}:
    get:
      tags:
      - Opds
      operationId: getOpdByPatient
      parameters:
      - name: pcode
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpdWithOperationRowDTO"
      security:
      - bearerAuth: []
...
  /laboratories/examRequest/patient/{patId}:
    get:
      tags:
      - Laboratories
      operationId: getLaboratoryExamRequest_1
      parameters:
      - name: patId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LaboratoryDTO"
      security:
      - bearerAuth: []
...
  /admissions/patient/{patientCode}:
    get:
      tags:
      - Admissions
      operationId: getAdmissions_1
      parameters:
      - name: patientCode
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "400":
          description: Bad Request
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdmissionDTO"
      security:
      - bearerAuth: []
...
  /admissions/current:
    get:
      tags:
      - Admissions
      operationId: getCurrentAdmission
      parameters:
      - name: patientCode
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "400":
          description: Bad Request
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionDTO"
      security:
      - bearerAuth: []
...
components:
  schemas:
    WardDTO:
      type: object
      properties:
        code:
          type: string
          description: The ward code
          maxLength: 3
        description:
          type: string
          description: The name of the ward
          maxLength: 50
        telephone:
          type: string
          description: The phone number of the ward
          maxLength: 50
        fax:
          type: string
          description: The fax number of the ward
          maxLength: 50
        email:
          type: string
          description: The email of the ward
          maxLength: 50
        beds:
          type: integer
          format: int32
        nurs:
          type: integer
          format: int32
        docs:
          type: integer
          format: int32
        visitDuration:
          type: integer
          format: int32
        lock:
          type: integer
          format: int32
          description: lock
          example: 0
        opd:
          type: boolean
        female:
          type: boolean
        male:
          type: boolean
        pharmacy:
          type: boolean
      required:
      - beds
      - description
      - docs
      - nurs
      - visitDuration
    PatientDTO:
      type: object
      description: Class representing a patient
      properties:
        code:
          type: integer
          format: int32
          readOnly: true
        firstName:
          type: string
          description: First name of the patient
          example: Mario
          maxLength: 50
        secondName:
          type: string
          description: Last name of the patient
          example: Rossi
          maxLength: 50
        name:
          type: string
          readOnly: true
        birthDate:
          type: string
          description: Birth date
          example: 1979-05-01
        age:
          type: integer
          format: int32
          description: Age
          example: 40
        agetype:
          type: string
          description: Age type
          example: null
        sex:
          type: string
          description: Sex
          enum:
          - M
          - F
          example: M
        address:
          type: string
          description: Address
          example: "Via Roma, 12"
          maxLength: 50
        city:
          type: string
          description: City
          example: Verona
          maxLength: 50
        nextKin:
          type: string
          readOnly: true
        telephone:
          type: string
          description: Telephone
          example: "+393456789012"
          maxLength: 50
        note:
          type: string
          description: Note
          example: Test insert new patient
          maxLength: 65535
        motherName:
          type: string
          description: Mother's name
          example: Roberta
          maxLength: 50
        mother:
          type: string
          description: "Mother's status (D=dead, A=alive)"
          enum:
          - D
          - A
          example: A
        fatherName:
          type: string
          description: Father's name
          example: Giuseppe
          maxLength: 50
        father:
          type: string
          description: "Father's status (D=dead, A=alive)"
          enum:
          - D
          - A
          example: D
        bloodType:
          type: string
          description: "Blood type (0-/+, A-/+ , B-/+, AB-/+)"
          enum:
          - 0-
          - 0+
          - A-
          - A+
          - B-
          - B+
          - AB-
          - AB+
          example: A+
        hasInsurance:
          type: string
          description: "HasInsurance (Y=Yes, N=no)"
          enum:
          - "Y"
          - "N"
          example: "N"
        parentTogether:
          type: string
          description: "Parent together (Y=Yes, N=no)"
          enum:
          - "Y"
          - "N"
          example: "N"
        taxCode:
          type: string
          description: Tax code
          example: RSSMRA79E01L781N
          maxLength: 30
        lock:
          type: integer
          format: int32
          description: Lock
          example: 0
        blobPhoto:
          type: string
          format: byte
          description: BlobPhoto
        hashCode:
          type: integer
          format: int32
          readOnly: true
        allergies:
          type: string
          description: Allergies of patient
          maxLength: 255
        anamnesis:
          type: string
          description: Current anamnesis
          maxLength: 255
        status:
          type: string
          description: Status
          enum:
          - I
          - O
          example: I
        consensusFlag:
          type: boolean
          description: Consensus flag
          example: true
        consensusServiceFlag:
          type: boolean
          description: Consensus service flag
          example: true
      required:
      - age
      - bloodType
      - city
      - fatherName
      - firstName
      - motherName
      - secondName
      - sex
    VisitDTO:
      type: object
      description: Class representing a vaccine type
      properties:
        visitID:
          type: integer
          format: int32
          description: The visit's ID
        patient:
          $ref: "#/components/schemas/PatientDTO"
          description: Patient related to visitor
        date:
          type: string
          description: Date of the visit
          example: 2020-03-19T14:58:00
        note:
          type: string
          description: Note of the visit
          maxLength: 65535
        sms:
          type: boolean
          description: Sms of the visit
        ward:
          $ref: "#/components/schemas/WardDTO"
          description: Ward of the visit
        duration:
          type: integer
          format: int32
          description: Duration of the visit
        service:
          type: string
          description: Service done during the visit
          maxLength: 45
        lock:
          type: integer
          format: int32
          description: Lock
          example: 0
      required:
      - date
      - patient
...
    PatientConsensusDTO:
      type: object
      description: Class representing a patient consensus
      properties:
        consensusFlag:
          type: boolean
          description: Consensus flag
        serviceFlag:
          type: boolean
          description: Service flag
        patientId:
          type: integer
          format: int32
          description: Patient id
      required:
      - patientId
...
    AdmissionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Admission key
          example: 12
        admitted:
          type: integer
          format: int32
          description: If admitted or not
          example: 0
        type:
          type: string
          description: Type of admission
          example: "N"
        ward:
          $ref: "#/components/schemas/WardDTO"
          description: Ward
        patient:
          $ref: "#/components/schemas/PatientDTO"
          description: Patient
        admDate:
          type: string
          description: Admission date
        admType:
          $ref: "#/components/schemas/AdmissionTypeDTO"
          description: Admission type
        diseaseIn:
          $ref: "#/components/schemas/DiseaseDTO"
          description: 'Disease in '
        diseaseOut1:
          $ref: "#/components/schemas/DiseaseDTO"
          description: 'Disease out '
        diseaseOut2:
          $ref: "#/components/schemas/DiseaseDTO"
          description: 'Disease out '
        diseaseOut3:
          $ref: "#/components/schemas/DiseaseDTO"
          description: 'Disease out '
        operation:
          $ref: "#/components/schemas/OperationDTO"
          description: 'Operation '
        opDate:
          type: string
          description: Operation date
        opResult:
          type: string
          description: 'Operation result value is ''P'' or ''N'' '
          example: "N"
        disDate:
          type: string
          description: Discharge date
        disType:
          $ref: "#/components/schemas/DischargeTypeDTO"
          description: DisChargeType
        note:
          type: string
          description: Free note
          maxLength: 65535
        transUnit:
          type: number
          format: float
          description: Transfusional unit
        visitDate:
          type: string
          description: Visit date
        pregTreatmentType:
          $ref: "#/components/schemas/PregnantTreatmentTypeDTO"
          description: 'TreatmentType '
        deliveryDate:
          type: string
          description: Delivery date
        deliveryType:
          $ref: "#/components/schemas/DeliveryTypeDTO"
          description: Delivery type
        deliveryResult:
          $ref: "#/components/schemas/DeliveryResultTypeDTO"
          description: Delivery result type
        weight:
          type: number
          format: float
          description: Weight
        ctrlDate1:
          type: string
          format: date-time
        ctrlDate2:
          type: string
          format: date-time
        abortDate:
          type: string
          format: date-time
        userID:
          type: string
          description: User id
        hashCode:
          type: integer
          format: int32
          readOnly: true
        lock:
          type: integer
          format: int32
          description: Lock
          example: 0
        deleted:
          type: string
          description: "Flag record deleted, values are 'Y' OR 'N' "
          example: "N"
        yprog:
          type: integer
          format: int32
        fhu:
          type: string
      required:
      - admDate
      - admitted
      - deleted
      - type
...
    OperationRowDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        operation:
          $ref: "#/components/schemas/OperationDTO"
          description: The operation
        prescriber:
          type: string
          description: The presciber of the operation
          maxLength: 150
        opResult:
          type: string
          description: The result of the operation
          maxLength: 250
        opDate:
          type: string
          description: Operation registration date
        remarks:
          type: string
          description: The remark of the operation
          maxLength: 250
        admission:
          $ref: "#/components/schemas/AdmissionDTO"
          description: The admission
        opd:
          $ref: "#/components/schemas/OpdDTO"
          description: The opd
        bill:
          $ref: "#/components/schemas/BillDTO"
          description: The bill
        transUnit:
          type: number
          format: float
          description: The transunit
        hashCode:
          type: integer
          format: int32
          readOnly: true
      required:
      - opDate
      - opResult
      - operation
      - prescriber
    OpdWithOperationRowDTO:
      type: object
      properties:
        opdDTO:
          $ref: "#/components/schemas/OpdDTO"
        operationRows:
          type: array
          items:
            $ref: "#/components/schemas/OperationRowDTO"
...
    ExamDTO:
      type: object
      properties:
        code:
          type: string
          description: Exam Code
          example: "01.01"
          maxLength: 10
        description:
          type: string
          description: Exam Description
          example: 1.1
          maxLength: 100
        procedure:
          type: integer
          format: int32
          description: Exam Procedure
          example: 1
        defaultResult:
          type: string
          description: Exam Default Result
          example: '>=12 (NORMAL)'
          maxLength: 50
        examtype:
          $ref: "#/components/schemas/ExamTypeDTO"
          description: Exam Type
        lock:
          type: integer
          format: int32
          description: Lock
          example: 0
    ExamTypeDTO:
      type: object
      properties:
        code:
          type: string
          description: Exam Type Code
          example: HB
        description:
          type: string
          description: Exam Type Description
          example: 1.Haematology
          maxLength: 50
...
    LaboratoryDTO:
      type: object
      properties:
        code:
          type: integer
          format: int32
          description: Laboratory Code
        material:
          type: string
          description: Laboratory Material
          example: Blood
        exam:
          $ref: "#/components/schemas/ExamDTO"
          description: Laboratory Exam
        registrationDate:
          type: string
          description: Laboratory Registration Date
        labDate:
          type: string
          description: Laboratory Exam Date
        result:
          type: string
          description: Laboratory Result
        lock:
          type: integer
          format: int32
          description: Lock
          example: 0
        note:
          type: string
          description: Laboratory Note
          example: Note by laboratorist
          maxLength: 255
        patientCode:
          type: integer
          format: int32
          description: Laboratory Patient Code
        patName:
          type: string
          description: Laboratory Patient Name
        inOutPatient:
          type: string
          description: Laboratory Patient InOut
          enum:
          - I
          - O
          example: 0
        age:
          type: integer
          format: int32
          description: Laboratory Patient Age
        sex:
          type: string
          description: Laboratory Patient Sex
          example: M
        status:
          type: string
          description: Laboratory status
          enum:
          - draft
          - open
          - done
          - invalid
          - deleted
          example: DRAFT
...
    PageInfoDTO:
      type: object
      properties:
        size:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        nbOfElements:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int64
        totalNbOfElements:
          type: integer
          format: int64
        hasPreviousPage:
          type: boolean
        hasNextPage:
          type: boolean
    PagePatientDTO:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PatientDTO"
        pageInfo:
          $ref: "#/components/schemas/PageInfoDTO"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT